Ecommerce/requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-allauth>=65.0.0
django-cors-headers>=4.3.0
django-filter>=24.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
razorpay>=1.4.2
twilio>=9.0.0
Pillow>=10.0.0
cryptography>=42.0.0
